zones tests migrated 1:1
