--- cloud/linode/zones.go
@@ -1,43 +1,0 @@
-package linode
-
-import (
-	"context"
-
-	"k8s.io/apimachinery/pkg/types"
-	cloudprovider "k8s.io/cloud-provider"
-)
-
-// TODO(PR): move this to instancesv2; move tests to instances_test
-type zones struct {
-	client Client
-	region string
-}
-
-func newZones(client Client, zone string) cloudprovider.Zones {
-	return zones{client, zone}
-}
-
-func (z zones) GetZone(_ context.Context) (cloudprovider.Zone, error) {
-	return cloudprovider.Zone{Region: z.region}, nil
-}
-
-func (z zones) GetZoneByProviderID(ctx context.Context, providerID string) (cloudprovider.Zone, error) {
-	id, err := parseProviderID(providerID)
-	if err != nil {
-		return cloudprovider.Zone{}, err
-	}
-	linode, err := linodeByID(ctx, z.client, id)
-	if err != nil {
-		return cloudprovider.Zone{}, err
-	}
-
-	return cloudprovider.Zone{Region: linode.Region}, nil
-}
-
-func (z zones) GetZoneByNodeName(ctx context.Context, nodeName types.NodeName) (cloudprovider.Zone, error) {
-	linode, err := linodeByName(ctx, z.client, nodeName)
-	if err != nil {
-		return cloudprovider.Zone{}, err
-	}
-	return cloudprovider.Zone{Region: linode.Region}, nil
-}
